requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
anthropic>=0.89.0
httpx>=0.27.0
python-dotenv>=1.0.0
sqlalchemy>=1.4.0,<2.0.0
pyyaml>=6.0.1
aiosqlite>=0.19.0
requests>=2.31.0
langgraph>=0.2.0
langsmith>=0.1.0
langchain-anthropic>=0.1.0
redis>=5.0.0
aioredis>=2.0.0
asyncpg>=0.30.0
